+++ src/CrawlObserver.php
@@ -3,33 +3,50 @@
 namespace Spatie\Crawler;
 
 use Psr\Http\Message\UriInterface;
+use Psr\Http\Message\ResponseInterface;
+use GuzzleHttp\Exception\RequestException;
 
-interface CrawlObserver
+abstract class CrawlObserver
 {
     /**
      * Called when the crawler will crawl the url.
      *
      * @param \Psr\Http\Message\UriInterface $url
-     *
-     * @return void
      */
-    public function willCrawl(UriInterface $url);
+    public function willCrawl(UriInterface $url)
+    {
+    }
 
     /**
-     * Called when the crawler has crawled the given url.
+     * Called when the crawler has crawled the given url successfully.
      *
      * @param \Psr\Http\Message\UriInterface $url
-     * @param \Psr\Http\Message\ResponseInterface|null $response
+     * @param \Psr\Http\Message\ResponseInterface $response
      * @param \Psr\Http\Message\UriInterface|null $foundOnUrl
-     *
-     * @return void
      */
-    public function hasBeenCrawled(UriInterface $url, $response, ?UriInterface $foundOnUrl = null);
+    abstract public function crawled(
+        UriInterface $url,
+        ResponseInterface $response,
+        ?UriInterface $foundOnUrl = null
+    );
+
+    /**
+     * Called when the crawler had a problem crawling the given url.
+     *
+     * @param \Psr\Http\Message\UriInterface $url
+     * @param \GuzzleHttp\Exception\RequestException $requestException
+     * @param \Psr\Http\Message\UriInterface|null $foundOnUrl
+     */
+    abstract public function crawlFailed(
+        UriInterface $url,
+        RequestException $requestException,
+        ?UriInterface $foundOnUrl = null
+    );
 
     /**
      * Called when the crawl has ended.
-     *
-     * @return void
      */
-    public function finishedCrawling();
+    public function finishedCrawling()
+    {
+    }
 }